fixes #51, #71, #72, #73,
- Fixed an issue where an exception would be thrown in non-outline mode
after attempting to use the talon view
- Fixed an issue where performing a mouse released would cause the card
movement to not be applied properly
- Fixed an issue where you could still drag the card after it was auto
placed within the foundation view
- Fixed an issue where single clicking on a card that has it's backside
shown would only uncover on `mouseReleased`, now it will be performed on
`mousePressed`.
- Fixed an issue where double clicking on a card would only attempt to
auto-position the card if on `mouseReleased`, now it will be done on
`mousePressed`.
- Fixed an issue where the stock view would only initiate a `next card`
call on `mouseReleased`. It will now instead perform the desired call on
`mousePressed`